style: better abstraction levels in scheduled messages
