+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/ConfirmDeleteMessageState.kt
@@ -1,9 +1,9 @@
 package com.github.heheteam.adminbot.states
 
 import com.github.heheteam.commonlib.EduPlatformError
+import com.github.heheteam.commonlib.ScheduledMessage
 import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.interfaces.AdminId
-import com.github.heheteam.commonlib.interfaces.ScheduledMessage
 import com.github.heheteam.commonlib.interfaces.ScheduledMessageId
 import com.github.heheteam.commonlib.state.BotStateWithHandlers
 import com.github.heheteam.commonlib.state.UpdateHandlerManager
