+++ admin-bot/src/main/kotlin/com/github/heheteam/adminbot/states/QueryFullTextConfirmationState.kt
@@ -2,10 +2,10 @@ package com.github.heheteam.adminbot.states
 
 import com.github.heheteam.commonlib.EduPlatformError
 import com.github.heheteam.commonlib.NamedError
+import com.github.heheteam.commonlib.ScheduledMessage
 import com.github.heheteam.commonlib.api.AdminApi
 import com.github.heheteam.commonlib.interfaces.AdminId
 import com.github.heheteam.commonlib.interfaces.CourseId
-import com.github.heheteam.commonlib.interfaces.ScheduledMessage
 import com.github.heheteam.commonlib.state.BotStateWithHandlers
 import com.github.heheteam.commonlib.state.UpdateHandlerManager
 import com.github.heheteam.commonlib.util.HandlingError
